Finally figured out the issue with the root pane not resizing
This fix is a total hack but it's just to show that the
CacioComponentPeer needs to resize its underlying Swing widget. Why
I didn't realize this before is anybody's guess.

The setBounds should really be done in response to ComponentEvents
in CacioComponentPeer and there should be a flag to stop the change
propogating back to the PlatformWindow. I'll post on cacio-dev to
make sure I know what I'm talking about.