Going to give lwawt a quick try
I can get this set up in a couple of hours and I'm already ahead
of schedule so I'm going to go for it. I think it might be more
suitable for desktop AWT ports rather than the embedded systems
that Cacio largely targets. Plus we get all the bugfixes from the
Mac OS X port which will be well supported.

--HG--
branch : lwawt