Fix build on Linux with 64kb default page size
There's a general problem with the specification of the shadow pages in HotSpot because they are specified in 'page' units (StackYellowPages, StackRedPages, StackShadowPages) but with the implicit assumption that the default system memory page size is 4kb. On Linux/ppc64 it is not unusual to have bigger (e.g. 64kb) page sizes and this leads to problems when HotSpot computes its minimum stack size. The general fix for this problem would be to change the specification of the shadow pages from 'page' units to bytes. But this would require shared code changes so for the time being we just increase the explicit stack size settings in the makefiles to not conflict with the minimum stack sizes computed by HotSpot.